Added getMember(int) method to Group to get the member at the specified index in the Group's member list.
Made DefaultTreeView tree virtual. All top-level items in the file are populated upon opening the file, the rest of the treeitems are created on-demand as groups are expanded.

Fixed graphical issue caused when expanding groups by disabling tree redraw until all the treeitems are done being created.

worked on issue where context menu isn't shown for data objects when they are clicked on for the first time and selectedObject is null